requirements.txt
appnope==0.1.0
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.1
cffi==1.14.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
imageio==2.8.0
import-ipynb==0.1.3
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numba==0.48.0
numpy==1.18.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
resampy==0.2.2
scikit-image==0.16.2
scikit-learn==0.22.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
SoundFile==0.10.3.post1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
